linkers, compilers: port remaining long->int32 fixes from the k variants
- lput/lputl/wput/wputl signatures in 5l/8l (from kencc commit 37e0fe69,
  which predates the principia import): no output change on gcc/Linux
  but needed to compile with clang on macOS
- elf.c/elf_.h (shared by 5l and 8l): Putl and the elf32 helpers follow,
  long/ulong -> int32/uint32, matching 8lk/elf.h
- elf32sectab: page-align the data section file offset like the phdr
  (rnd() ported from 8lk); only visible in 8l -S output since only -S
  emits the section table
- last (long) casts in the two remaining ieeedtod copies (cc2/pswt.c,
  assemblers/as/float.c), parity with cck/lexbody

This exhausts the long->int32 deltas between the lineages; the remaining
longs (Prog pc, Sym value, ...) are identical in both.

+++ compilers/cc2/pswt.c
@@ -214,7 +214,7 @@ ieeedtod(Ieee *ieee, double native)
     fr = modf(fr*f, &ho);
     ieee->l = ho;
     ieee->l <<= 16;
-    ieee->l |= (long)(fr*f);
+    ieee->l |= (int32)(fr*f);
 }
 /*e: function [[ieeedtod]] */
 /*e: cc2/pswt.c */
